Move rancher controller below seed-controller-manager pkg
